Create H3 Index

Create H3 indices from sfc objects, vectors of x and y coordinates, or H3 string IDs.

h3_from_xy(x, y, resolution)

h3_from_points(x, resolution)

h3_from_strings(x)

h3_to_points(x)

h3_to_vertexes(x)

# S3 method for H3
as.character(x)

Arguments

x

for h3_from_points() an object of class sfc_POINT. For h3_from_strings() a character vector of H3 index IDs. For h3_from_xy() a numeric vector of longitudes.

y

a numeric vector of latitudes.

resolution

an integer indicating the H3 cell resolution. Must be between 0 and 15 inclusive.

Details

h3_from_points(): takes an sfc_POINT object and creates a vector of H3 cells

h3_from_strings(): converts a character vector of cell indexes to an H3 vector

h3_from_xy(): converts vectors of x and y coordinates to an H3 vector

h3_to_points(): converts an H3 vector to a either an sfc_POINT object or a list of sfg POINT objects.

h3_to_vertexes(): converts an H3 vector to an sfc_MULTIPOINT object or a list of MULTIPOINT objects.

Examples

h3_from_xy(-90, 120, 5)
#> <H3[1]>
#> [1] 85f29383fffffff

h3_from_strings("85f29383fffffff")
#> <H3[1]>
#> [1] 85f29383fffffff

if (requireNamespace("sf")) {
  # create random points
  pnts <- sf::st_cast(
    sf::st_sfc(
      sf::st_multipoint(matrix(runif(10, max = 90), ncol = 2)),
      crs = 4326
    ),
    "POINT"
  )

  # convert to H3 objects
  h3s <- h3_from_points(pnts, 5)

  h3_to_vertexes(h3s)

  h3_to_points(h3s)
}
#> Geometry set for 5 features 
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 24.72192 ymin: 14.94596 xmax: 73.52265 ymax: 42.13182
#> Geodetic CRS:  WGS 84
#> POINT (24.72192 37.3928)
#> POINT (73.52265 23.07089)
#> POINT (45.27477 14.94596)
#> POINT (46.3641 16.37337)
#> POINT (27.81511 42.13182)